--- src/model/faster_rcnn.py
@@ -1,20 +1,0 @@
-import torchvision
-from torchvision.models.detection import FasterRCNN
-from torchvision.models.detection.rpn import AnchorGenerator
-
-
-def get_model_instance_segmentation(num_classes):
-    backbone = torchvision.models.mobilenet_v2(pretrained=True).features
-    backbone.out_channels = 1280
-    anchor_generator = AnchorGenerator(
-        sizes=((8, 16, 32, 64, 128),),
-        aspect_ratios=((0.5, 1.0, 2.0),)
-    )
-
-    model = FasterRCNN(
-        backbone,
-        num_classes=num_classes,
-        rpn_anchor_generator=anchor_generator
-    )
-
-    return model
